the great int-ening

+++ assignment3/src/Bulbasaur.java
@@ -1,5 +1,5 @@
 public class Bulbasaur extends Pokemon { 
-    public Bulbasaur(short level) {
+    public Bulbasaur(int level) {
         super(level, 0.20, "Bulbasaur");
     }
 }
