+++ assignment3/src/Caterpie.java
@@ -1,5 +1,5 @@
 public class Caterpie extends Pokemon {
-    public Caterpie(short level) {
+    public Caterpie(int level) {
         super(level, 0.5, "Caterpie");
     }
 }
